Use a better naming for the modules and re-export the required modules
