--- src/atom/mod.rs
@@ -1,14 +1,0 @@
-/*
- * © 2022 Arastoo Bozorgi
- * All rights reserved.
- */
-
-
-pub mod movie;
-pub mod ftype;
-pub mod media_data;
-pub mod mp4_atom;
-pub mod free;
-pub mod track;
-pub mod inner;
-pub mod media;
