+++ src/boxes/media/sample_table.rs
@@ -3,38 +3,23 @@
  * All rights reserved.
  */
 
-use crate::atom::mp4_atom::{Mp4Atom, AtomName};
+use crate::boxes::{Mp4Atom, AtomName, InnerAtom};
 use crate::error::Error;
 use byteorder::{BigEndian, ByteOrder};
-use crate::atom::free::Free;
-use crate::atom::track::Track;
-use crate::header::Header;
-use crate::atom::inner::InnerAtom;
+use crate::Header;
 
-pub struct Movie {
+pub struct SampleTable {
     atoms: Vec<Box<dyn Mp4Atom>>,
     header: Header
 }
 
-// impl Movie {
-//
-// }
-
-struct MovieHeader {
-    header: Header,
-    data: Vec<u8>
+struct SampleDesc {
+    atoms: Vec<Box<dyn Mp4Atom>>,
+    header: Header
 }
 
-struct Udata {
-    header: Header,
-    data: Vec<u8>
-}
-
-
-
-impl Mp4Atom for Movie {
-
-    fn parse(data: &[u8], start: usize) -> Result<Self, Error> {
+impl Mp4Atom for SampleTable {
+    fn parse(data: &[u8], start: usize) -> Result<Self, Error> where Self: Sized {
         let mut atoms = vec![];
         let header = Header::header(data, start)?;
         let mut index = 8; // skip the first 8 bytes that are Movie headers
@@ -48,30 +33,12 @@ impl Mp4Atom for Movie {
             let name = AtomName::from(name);
 
             let atom = match name {
-                AtomName::MovieHeader => {
+                AtomName::SampleDesc => {
                     Box::new(
-                        MovieHeader::parse(&data[index..index + size], index + start)?
-                    )
-                        as Box<dyn Mp4Atom>
-                }
-                AtomName::Track => {
-                    Box::new(
-                        Track::parse(&data[index..index + size], index + start)?
+                        SampleDesc::parse(&data[index..index + size], index + start)?
                     )
                         as Box<dyn Mp4Atom>
                 },
-                AtomName::Udata => {
-                    Box::new(
-                        Udata::parse(&data[index..index + size], index + start)?
-                    )
-                        as Box<dyn Mp4Atom>
-                },
-                AtomName::Free =>  {
-                    Box::new(
-                        Free::parse(&data[index..index + size], index + start)?
-                    )
-                        as Box<dyn Mp4Atom>
-                }
                 _ => {
                     Box::new(
                         InnerAtom::parse(&data[index..index + size], index + start)?
@@ -115,12 +82,31 @@ impl Mp4Atom for Movie {
     }
 }
 
-impl Mp4Atom for MovieHeader {
-    fn parse(data: &[u8], start: usize) -> Result<Self, Error> {
-        let header = Header::header(&data, start)?;
-        Ok(MovieHeader {
-            header,
-            data: data.to_vec()
+
+impl Mp4Atom for SampleDesc {
+    fn parse(data: &[u8], start: usize) -> Result<Self, Error> where Self: Sized {
+        let mut atoms = vec![];
+        let header = Header::header(data, start)?;
+
+        // let mut index = 0;
+        // while index < data.len() {
+        //
+        //     // the first 8 bytes includes the atom size and its name
+        //     // The size is the entire size of the box, including the size and type header, fields, and all contained boxes.
+        //     let size = BigEndian::read_u32(&data[index..index + 4]) as usize;
+        //     let name = std::str::from_utf8(&data[index + 4..index + 8])?;
+        //
+        //     let atom = Box::new(
+        //         InnerAtom::parse(&data[index..index + size], index + start)?
+        //     ) as Box<dyn Mp4Atom>;
+        //
+        //     atoms.push(atom);
+        //     index += size;
+        // }
+
+        Ok(Self {
+            atoms,
+            header
         })
     }
 
@@ -145,40 +131,6 @@ impl Mp4Atom for MovieHeader {
     }
 
     fn internals(&self) -> Option<&Vec<Box<dyn Mp4Atom>>> {
-        None
-    }
-}
-
-impl Mp4Atom for Udata {
-    fn parse(data: &[u8], start: usize) -> Result<Self, Error> {
-        let header = Header::header(&data, start)?;
-        Ok(Udata {
-            header,
-            data: data.to_vec()
-        })
-    }
-
-    fn start(&self) -> usize {
-        self.header.start
-    }
-
-    fn end(&self) -> usize {
-        self.header.start + self.header.size
-    }
-
-    fn size(&self) -> usize {
-        self.header.size
-    }
-
-    fn name(&self) -> &str {
-        self.header.name.as_ref()
-    }
-
-    fn read(&self) -> Result<Vec<u8>, Error> {
-        unimplemented!()
-    }
-
-    fn internals(&self) -> Option<&Vec<Box<dyn Mp4Atom>>> {
-        None
+        Some(&self.atoms)
     }
 }
